Implementing token price statistic for listed tokens.

+++ Models/TokenContract.cs
@@ -10,7 +10,7 @@ namespace SYNCWallet.Models
     public class TokenContract
     {
         public int Network { get; set; }
-        public decimal Decimals { get; set; }
+        public int Decimals { get; set; }
         public string ContractAddress { get; set; }
         public decimal UserBalance { get; set; }
         public decimal Price { get; set; }
